fix for hubris on my part thinking that everyone would be wanting to run Chainsaw.
It is now not required to have Jakarta ORO to build just the Log4j core.

The 'build' target now does NOT compile any chainsaw files.

The 'chainsaw' target relies on the 'build' target and requires the ORO
configured otherwise an friendly error message is displayed.

Thanks to Raymond DeCampo for spotting this.