Fixed NPE bug found by Ulrich.Kleeberger@witt-weiden.de when attempting to close the welcome panel using the right click popup menu on the tabs.
Also reordered the position of this popup code so the L&F changes
affected the popup menu (was staying in the Metal L&F).

Now listens for tabs being removed, and disables the hide actions if the 
Welcome panel has been selected.

Tutorial Frame toolbar now tracks whether there are any Tutorial Generator
Receivers running, and if not, ensures that the Start Tutorial button
does not stay activated.