[This change should be reviewed, particularly by MarkW]
Change to fix the PluginTestCase.

The change from the use of the equals(...) method to
use the new isEquivalent(Plugin) method seems to have
missed the comparison check of the class of each
Plugin.  I have modified this to be inline with what the
existing Witness file is expecting.

If the class of the plugin should not form part of the
equivalence check, then the witness file will need to be modified.