beginnings of some work to improve usability of the focus on, ignore logger. within the Logger name tree panel.
If you hold down CTRL the mouse icon changes to a Focus On icon.  
This will eventually be hooked up so that while holding down CTRL
and double clicking you can focus on a logger.

Holding down CTRL+SHIFT will be the ignore action.

This mimics the way Photoshop's zoom works.