[Please review this change, particularly the synchronization semantics]
This set of changes removes the dependancy on the EventListenerList
class which seems to be JDK1.3.

This should allow the non-chainsaw areas of log4j to be compiled under JDK 1.2,
although I do not have this installed, and cannot verify.

All the test cases (bar the Scheduler one, but I don't believe
that is related to this change) pass after this change.