Bug 41156 Give log4j threads reasonable names.
Thread names are based on the class name that is the creator/related entity, and
where threads are based on Socket management, the port # being managed is included
in the name.  This follows a convention seen in many other packages (such as Tomcat)
to easily identify the resource the Thread is controlling.

I have renamed the AsyncAppender thread name that was already given an explicit name
to follow the above convention.